+++ src/Poe/FileInMemory.cs
@@ -5,23 +5,23 @@ namespace PoeHUD.Poe
 {
     public abstract class FileInMemory
     {
-        protected FileInMemory(Memory m, int address)
+        protected FileInMemory(Memory m, long address)
         {
             M = m;
             Address = address;
         }
 
         public Memory M { get; }
-        public int Address { get; }
+        public long Address { get; }
         private int NumberOfRecords => M.ReadInt(Address + 0x30, 0x20);
 
-        protected IEnumerable<int> RecordAddresses()
+        protected IEnumerable<long> RecordAddresses()
         {
-            int firstRec = M.ReadInt(Address + 0x30, 0x4);
-            int lastRec = M.ReadInt(Address + 0x30, 0x8);
+            long firstRec = M.ReadLong(Address + 0x30, 0x4);
+            long lastRec = M.ReadLong(Address + 0x30, 0x8);
             int cnt = NumberOfRecords;
 
-            int recLen = (lastRec - firstRec) / cnt;
+            long recLen = (lastRec - firstRec) / cnt;
             for (int i = 0; i < cnt; i++)
                 yield return firstRec + i * recLen;
         }
